ready room mouse fix. Scenes named differently depending on what the load sequence was
