+++ Assets/Scripts/GameConstants.cs
@@ -45,4 +45,12 @@ namespace Filibusters
         public static readonly string DEPOSIT = "Deposit";
         public static readonly string INACTIVE_OVERLAY = "InactiveIndicator";
     }
+
+    public class Scenes
+    {
+        public static readonly string START_MENU = "Scenes/StartMenu";
+        public static readonly string READY_MENU = "Scenes/ReadyMenu";
+        public static readonly string MAIN = "Scenes/Main";
+        public static readonly string GAME_OVER = "Scenes/GameOver";
+    }
 }
