--- Assets/Scripts/Scenes.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-namespace Filibusters
-{
-    public class Scenes
-    {
-        public static readonly string START_MENU = "Scenes/StartMenu";
-        public static readonly string READY_MENU = "Scenes/ReadyMenu";
-        public static readonly string MAIN = "Scenes/Main";
-        public static readonly string GAME_OVER = "Scenes/GameOver";
-    }
-}
