--- Assets/Scripts/Scenes.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0894d3641373f73448751c7d5689cdb0
-timeCreated: 1475462261
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
